Desenvolvimento de mappers, controller, serviços de consulta, classes e métodos utilitários, melhoria na POO, e aplicação de captura de logs.

+++ backend/src/main/java/com/personagens/rickymorty/repository/EpisodeRepository.java
@@ -1,9 +1,9 @@
 package com.personagens.rickymorty.repository;
 
-import entity.EpisodeEntity;
+import com.personagens.rickymorty.entity.EpisodeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EpisodeRepository extends JpaRepository<EpisodeEntity, Long> {
+public interface EpisodeRepository extends JpaRepository<EpisodeEntity, Integer> {
 }
